jsonGet.py: Simplify with EAFP instead of LBYL.
https://docs.python.org/3/glossary.html#term-eafp
https://docs.python.org/3/glossary.html#term-lbyl

In many languages, exception handing is slow, or painful,
so one tries to avoid exceptions by using LBYL technique.

Python excels at exception handling and relies on it heavily,
such as with the StopIteration exception for the iterator protocol.
Exception handling in Python is fast and easy to code,
so exception handling is not something to shun in Python.

from The Elements of Programming Style, 2nd edition 1978
by Kernighan and Plauger

    Use the good features of a language; avoid the bad ones.
    Use the "telephone test" for readability.

The new way is also DRYer; it only has one return default statement,
not three.

By the way, this commit changes behavior by allowing negative indexes into
lists. Is allowing negative indexes a feature or a bug?
I lean toward thinking it is a feature. What do you think?

Also, this commit changes behavior by allowing data that is not an instance of
a dict or list, but can nonetheless accept a key or index. Is that a bug or a
feature? I lean toward thinking it is a feature. What do you think?