jsonGet.py: Eliminate first_key is None test.
I had the feeling that it would be legitimate for first_key to be None,
so I deleted the test and return default.

from The Elements of Programming Style, 2nd edition 1978
by Kernighan and Plauger

    Make sure your code does "nothing" gracefully.

This is a speculative change that might have introduced a bug.

John will have to scrutinize this change.