jsonGet.py: Combined first_key and other_keys arguments into keys argument.
This is another speculative change that might be a bug.
The inspiration for this comes from
The Elements of Programming Style, 2nd edition 1978
by Kernighan and Plauger

    Make sure special cases are truly special.

I could not think of a reason why having no keys would be a bug,
so I eliminated the need for at least one key.

John will have to scrutinize this change.